classifier/requirements.txt

numpy
librosa
matplotlib
soundfile
pillow
